--- cypress/e2e/API_TESTING/pets/findPetsByTags.cy.js
@@ -1,75 +1,0 @@
-import Ajv from 'ajv';
-const ajv = new Ajv();
-const apiBaseURL = Cypress.env('CYPRESS_BASE_URL');
-let requestInfo = JSON.parse(
-    JSON.stringify({ url: '/pet/findByTags', method: 'GET' })
-);
-requestInfo.url = apiBaseURL + requestInfo.url;
-
-describe('Finds Pets by tags', () => {
-    // Please be informed that the API in this file has been deprecated and will no longer be supported.
-    it.skip('Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.', () => {
-        cy.fixture('200_application_json__findPetsByTags').then(
-            (fixtureResponse) => {
-                requestInfo.body = fixtureResponse.payload
-                    ? fixtureResponse.payload
-                    : '';
-                requestInfo.headers = fixtureResponse.headers
-                    ? fixtureResponse.headers
-                    : '';
-
-                requestInfo.qs = fixtureResponse.queryParam
-                    ? fixtureResponse.queryParam
-                    : '';
-
-                cy.request(requestInfo).then((response) => {
-                    expect(response.status).to.eq(
-                        parseInt(fixtureResponse.responseStatusCode)
-                    );
-                    if (
-                        fixtureResponse.responseSchema &&
-                        fixtureResponse.responseSchema != ''
-                    ) {
-                        const validate = ajv.compile(
-                            fixtureResponse.responseSchema
-                        );
-                        const isValid = validate(response.body);
-                        expect(isValid).to.be.true;
-                    }
-                });
-            }
-        );
-    });
-    it.skip('Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.', () => {
-        cy.fixture('200_application_xml__findPetsByTags').then(
-            (fixtureResponse) => {
-                requestInfo.body = fixtureResponse.payload
-                    ? fixtureResponse.payload
-                    : '';
-                requestInfo.headers = fixtureResponse.headers
-                    ? fixtureResponse.headers
-                    : '';
-
-                requestInfo.qs = fixtureResponse.queryParam
-                    ? fixtureResponse.queryParam
-                    : '';
-
-                cy.request(requestInfo).then((response) => {
-                    expect(response.status).to.eq(
-                        parseInt(fixtureResponse.responseStatusCode)
-                    );
-                    if (
-                        fixtureResponse.responseSchema &&
-                        fixtureResponse.responseSchema != ''
-                    ) {
-                        const validate = ajv.compile(
-                            fixtureResponse.responseSchema
-                        );
-                        const isValid = validate(response.body);
-                        expect(isValid).to.be.true;
-                    }
-                });
-            }
-        );
-    });
-});
