+++ cypress/fixtures/400_application_json__findPetsByStatus.json
@@ -1,0 +1,28 @@
+{
+    "headers": {
+        "accept": "application/json",
+        "header": ""
+    },
+    "responseStatusCode": "400",
+    "responseSchema": {
+        "type": "array",
+        "items": {
+            "type": "object",
+            "properties": {
+                "code": {
+                    "type": "integer",
+                    "format": "int32"
+                },
+                "type": {
+                    "type": "string"
+                },
+                "message": {
+                    "type": "string"
+                }
+            }
+        }
+    },
+    "queryParam": {
+        "undefined": ""
+    }
+}
